Refactored parse.py into module
