+++ bin/pipeline_result_processor/prp/models/sample.py
@@ -6,8 +6,8 @@ from pydantic import BaseModel, Field
 
 from .base import RWModel
 from .metadata import RunMetadata
-from .qc import QcMethodIndex
 from .phenotype import PhenotypeResult, PhenotypeType
+from .qc import QcMethodIndex
 from .typing import TypingMethod, TypingResultCgMlst, TypingResultMlst
 
 SAMPLE_ID_PATTERN = r"^[a-zA-Z1-9-_]+$"
