+++ bin/pipeline_result_processor/setup.py
@@ -1,5 +1,6 @@
-from setuptools import find_packages, setup
 import os
+
+from setuptools import find_packages, setup
 
 with open("README.md") as f:
     long_description = f.read()
